management/requirements.txt
Django==4.2.6
djangorestframework>=3.12.4,<3.13
Pillow>=8.2.0,<8.3.0